Refactor Persistent Workers
There are issues with the current implementation of persistent workers where an OOM will break the RxJava implementation and the message will not relay back to the Bazel server. This will cause builds to run indefinitely.

There is a stable version in Rules Kotlin. We will piggy back off of the public implementation.